Tiny refactor of test dir
