+++ test/services/getCourtCase.integration.test.ts
@@ -4,7 +4,7 @@ import CourtCase from "../../src/services/entities/CourtCase"
 import getDataSource from "../../src/services/getDataSource"
 import getCourtCase from "../../src/services/getCourtCase"
 import { isError } from "../../src/types/Result"
-import deleteFromTable from "../testFixtures/database/deleteFromTable"
+import deleteFromTable from "../util/deleteFromTable"
 import { getDummyCourtCase, insertCourtCases } from "../testFixtures/database/insertCourtCases"
 
 describe("getCourtCases", () => {
