+++ test/services/listCourtCases.integration.test.ts
@@ -4,7 +4,7 @@ import { expect } from "@jest/globals"
 import { DataSource } from "typeorm"
 import listCourtCases from "../../src/services/listCourtCases"
 import { ListCourtCaseResult } from "types/ListCourtCasesResult"
-import deleteFromTable from "../testFixtures/database/deleteFromTable"
+import deleteFromTable from "../util/deleteFromTable"
 import {
   insertCourtCasesWithCourtDates,
   insertCourtCasesWithCourtNames,
